Addded CartController, ControllersTests and more stuff

+++ BasicSupermarket/Dtos/Product/UpdateProductRequestDto.cs
@@ -10,6 +10,6 @@ public record UpdateProductRequestDto
     public string Description { get; set; }
     [Url]
     public string ImageUrl { get; set; }
-    public long Price { get; set; }
+    public decimal Price { get; set; }
     public int? CategoryId { get; set; }
 }
